+++ changelog.md
@@ -1,3 +1,9 @@
+## 0.2.9
+
+ - The `fn` argument in the `.transform` method is now optional.
+ - Transducers support via `.transduce` method.
+
+
 ## 0.2.8
 
 Method `.flatMapWithConcurrencyLimit` renamed to `.flatMapConcurLimit`.
@@ -7,6 +13,7 @@ New methods:
  - .transform
  - .timestamp
  - Kefir.bus
+
 
 ## 0.2.7